Steganography_tool_project/requirements.txt
opencv-python>=4.5.5
numpy>=1.22.0
pycryptodome>=3.15.0
pillow>=9.0.0
pyganme